requirements.txt
Flask==2.0.1
streamlit
tensorflow==2.15.0
flask-cors
requests
Werkzeug
numpy
pillow
pymongo
streamlit
extra_streamlit_components